+++ packages/amplify-category-api/src/index.ts
@@ -4,12 +4,16 @@ import path from 'path';
 import { run } from './commands/api/console';
 import { getCfnApiArtifactHandler } from './provider-utils/awscloudformation/cfn-api-artifact-handler';
 
-export { NETWORK_STACK_LOGICAL_ID } from './category-constants'
+export { NETWORK_STACK_LOGICAL_ID } from './category-constants';
 export { DEPLOYMENT_MECHANISM } from './provider-utils/awscloudformation/base-api-stack';
 export { EcsStack } from './provider-utils/awscloudformation/ecs-apigw-stack';
 export { EcsAlbStack } from './provider-utils/awscloudformation/ecs-alb-stack';
 export { getGitHubOwnerRepoFromPath } from './provider-utils/awscloudformation/utils/github';
-export { generateContainersArtifacts, ApiResource, processDockerConfig } from './provider-utils/awscloudformation/utils/containers-artifacts';
+export {
+  generateContainersArtifacts,
+  ApiResource,
+  processDockerConfig,
+} from './provider-utils/awscloudformation/utils/containers-artifacts';
 export { getContainers } from './provider-utils/awscloudformation/docker-compose';
 
 const category = 'api';