+++ packages/amplify-category-auth/src/provider-utils/awscloudformation/constants.ts
@@ -3,6 +3,7 @@ import * as path from 'path';
 // path constants
 export const resourcesRoot = path.normalize(path.join(__dirname, '../../../resources'));
 export const adminAuthAssetRoot = path.join(resourcesRoot, 'adminAuth');
+export const authTriggerAssetFilePath = path.join(resourcesRoot, 'lambda-function.js');
 export const cfnTemplateRoot = path.join(resourcesRoot, 'cloudformation-templates');
 export const triggerRoot = path.normalize(path.join(__dirname, '../../../provider-utils/awscloudformation/triggers'));