+++ packages/amplify-container-hosting/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.6](https://github.com/aws-amplify/amplify-cli/compare/amplify-container-hosting@1.3.4...amplify-container-hosting@1.3.6) (2021-05-03)
+
+
+
+## 4.50.1 (2021-05-03)
+
+**Note:** Version bump only for package amplify-container-hosting
+
+
+
+
+
+## [1.3.5](https://github.com/aws-amplify/amplify-cli/compare/amplify-container-hosting@1.3.4...amplify-container-hosting@1.3.5) (2021-05-03)
+
+**Note:** Version bump only for package amplify-container-hosting
+
+
+
+
+
 ## [1.3.4](https://github.com/aws-amplify/amplify-cli/compare/amplify-container-hosting@1.3.3...amplify-container-hosting@1.3.4) (2021-04-27)
 
 **Note:** Version bump only for package amplify-container-hosting