+++ packages/amplify-nodejs-function-template-provider/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.6.6](https://github.com/aws-amplify/amplify-cli/compare/amplify-nodejs-function-template-provider@1.6.4...amplify-nodejs-function-template-provider@1.6.6) (2021-05-03)
+
+
+
+## 4.50.1 (2021-05-03)
+
+**Note:** Version bump only for package amplify-nodejs-function-template-provider
+
+
+
+
+
+## [1.6.5](https://github.com/aws-amplify/amplify-cli/compare/amplify-nodejs-function-template-provider@1.6.4...amplify-nodejs-function-template-provider@1.6.5) (2021-05-03)
+
+**Note:** Version bump only for package amplify-nodejs-function-template-provider
+
+
+
+
+
 ## [1.6.4](https://github.com/aws-amplify/amplify-cli/compare/amplify-nodejs-function-template-provider@1.6.3...amplify-nodejs-function-template-provider@1.6.4) (2021-04-27)
 
 **Note:** Version bump only for package amplify-nodejs-function-template-provider